+++ src/youwol/app/routers/environment/router.py
@@ -124,9 +124,13 @@ class YouwolEnvironmentResponse(BaseModel):
         proxied_backends = [
             ProxyInfo(
                 name=proxy.name,
+                uid=proxy.uid,
+                partitionId=proxy.partition_id,
                 version=proxy.version,
                 port=proxy.port,
+                startedAt=proxy.started_at,
                 pid=proxy.process and proxy.process.pid,
+                configuration=proxy.configuration,
             )
             for proxy in yw_env.proxied_backends.store
         ]