+++ CHANGELOG.md
@@ -18,6 +18,17 @@ and this project adheres to [PEP 440 Versioning](https://peps.python.org/pep-044
 
 ## [0.1.12.dev] − Unreleased
 
+### Added
+
+- **Custom Backends**:
+  - Custom backends now run within isolated partitions and can be dynamically configured during the build stage.
+    <!-- TG-2389 -->
+
+### Changed
+
+- Upgrade `GET:/co-lab` target from `@youwol/co-lab#^0.3.0` to `@youwol/co-lab#^0.4.0`. <!-- TG-2399 -->
+
+
 ## [0.1.11] − 2024-06-20
 
 ### Changed